+++ tapo/src/responses.rs
@@ -14,6 +14,7 @@ mod device_usage_energy_monitoring_result;
 mod device_usage_result;
 mod energy_data_result;
 mod energy_usage_result;
+mod next_event_result;
 mod power_data_result;
 mod preset;
 mod rtsp_stream_url;
@@ -40,6 +41,7 @@ pub use device_usage_energy_monitoring_result::*;
 pub use device_usage_result::*;
 pub use energy_data_result::*;
 pub use energy_usage_result::*;
+pub use next_event_result::NextEvent;
 pub use power_data_result::*;
 pub use preset::*;
 pub use rtsp_stream_url::*;
@@ -49,6 +51,7 @@ pub use trigger_logs_result::*;
 
 pub(crate) use control_child_result::*;
 pub(crate) use decodable_result_ext::*;
+pub(crate) use next_event_result::RawNextEvent;
 pub(crate) use schedule_rule_result::*;
 #[cfg(feature = "debug")]
 pub(crate) use supported_alarm_type_list_result::*;